Refactor repository interfaces and implementations to use size_t for count methods
- Updated count methods in various repository interfaces to return size_t instead of long for better type consistency and to reflect the expected size of collections.
- Modified corresponding implementations in memory persistence classes to align with the updated interface definitions.
- This change enhances type safety and improves clarity regarding the expected return type for count operations across the codebase.

+++ html-repository/source/uim/platform/html_repository/application/use_cases/manage_app_files.d
@@ -81,7 +81,7 @@ class ManageAppFilesUseCase : UIMUseCase {
         repo.remove(id);
     }
 
-    long countByVersion(AppVersionId versionId) {
+    size_t countByVersion(AppVersionId versionId) {
         return repo.countByVersion(versionId);
     }
