+++ html-repository/source/uim/platform/html_repository/infrastructure/persistence/memory/content_cache_repository.d
@@ -79,8 +79,8 @@ class ContentCacheMemoryRepository : ContentCacheRepository {
     store = result;
   }
 
-  long countByTenant(TenantId tenantId) {
-    long count = 0;
+  size_t countByTenant(TenantId tenantId) {
+    size_t count = 0;
     foreach (e; store) {
       if (e.tenantId == tenantId) count++;
     }
