+++ identity-provisioning/source/uim/platform/identity/provisioning/domain/ports/repositories/provisioning_log.d
@@ -14,8 +14,8 @@ interface ProvisioningLogRepository {
   ProvisioningLog[] findByJob(ProvisioningJobId jobtenantId, id tenantId);
   ProvisioningLog[] findByEntity(string entitytenantId, id tenantId);
   ProvisioningLog[] findByStatus(TenantId tenantId, LogStatus status);
-  long countByJob(ProvisioningJobId jobtenantId, id tenantId);
-  long countByJobAndStatus(ProvisioningJobId jobtenantId, id tenantId, LogStatus status);
+  size_t countByJob(ProvisioningJobId jobtenantId, id tenantId);
+  size_t countByJobAndStatus(ProvisioningJobId jobtenantId, id tenantId, LogStatus status);
   void save(ProvisioningLog entity);
   void remove(ProvisioningLogId tenantId, id tenantId);
   void removeByJob(ProvisioningJobId jobtenantId, id tenantId);
