+++ logging/source/uim/platform/logging/infrastructure/persistence/memory/notification_channel_repo.d
@@ -46,8 +46,8 @@ class MemoryNotificationChannelRepository : NotificationChannelRepository {
     store.remove(id);
   }
 
-  long countByTenant(TenantId tenantId) {
-    long count;
+  size_t countByTenant(TenantId tenantId) {
+    size_t count;
     foreach (ref ch; store)
       if (ch.tenantId == tenantId)
         count++;
