+++ mobile/source/uim/platform/mobile/application/use_cases/manage_app_versions.d
@@ -71,7 +71,7 @@ class ManageAppVersionsUseCase : UIMUseCase {
         repo.remove(id);
     }
 
-    long countByApp(MobileAppId appId) {
+    size_t countByApp(MobileAppId appId) {
         return repo.countByApp(appId);
     }
