+++ mobile/source/uim/platform/mobile/application/use_cases/manage_push_notifications.d
@@ -60,7 +60,7 @@ class ManagePushNotificationsUseCase : UIMUseCase {
         repo.remove(id);
     }
 
-    long countByApp(MobileAppId appId) {
+    size_t countByApp(MobileAppId appId) {
         return repo.countByApp(appId);
     }
